Move app.js da primeira aula para outro arquivo, criando um novo app.js para a API de produtos.
